Colour_Detection/requirements.txt
opencv-python==4.6.0.66
numpy>=1.26.0
Pillow==9.2.0